refactor(handy-core): move resampler & visualizer
