+++ crates/handy-core/src/audio/mod.rs
@@ -1,4 +1,6 @@
 pub mod constants;
+pub mod resampler;
 pub mod utils;
+pub mod visualizer;
 
 pub use constants::*;
